requirements.txt
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
joblib==1.3.2
numpy==1.26.3
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
pycparser==2.21
pyhtml2pdf==0.0.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
rapidfuzz==3.6.1
regex==2023.12.25
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
selenium==4.17.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
threadpoolctl==3.2.0
tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
tzdata==2023.4
urllib3==2.1.0
webdriver-manager==4.0.1
wsproto==1.2.0
